+++ tremor-cli/tests/integration/udp-reuseport/after/stop.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+if test -f before.pid; then
+    kill -QUIT $( cat before.pid )
+    # Remove pid file
+    rm -f before.pid
+fi